[Android Viewer] Fix crash after save dialog
Fixed the crash after save dialog caused by
a possible null pointer related to presentation
mode.

Reviewed-on: https://gerrit.libreoffice.org/39847
Tested-by: Jenkins <ci@libreoffice.org>

+++ android/source/src/java/org/libreoffice/LibreOfficeMainActivity.java
@@ -414,7 +414,7 @@ public class LibreOfficeMainActivity extends AppCompatActivity implements Settin
                 // noinspection ResultOfMethodCallIgnored
                 mTempFile.delete();
             }
-            if (mTempSlideShowFile.exists()) {
+            if (mTempSlideShowFile != null && mTempSlideShowFile.exists()) {
                 // noinspection ResultOfMethodCallIgnored
                 mTempSlideShowFile.delete();
             }
